Further tweaks to bass
Turns out gs :m inverted is lovely. D# not in our scale but shhhhhh.  Also a nice bass instrument stringing together some synths. Fm for a hard hit on start, dsaw for the floaty ending and sliced with varying octave to add pace. Finally a touch of darkseahorn to add a little grit and instability to the mix. Tends to our shine through on the longer notes. Still todo: work out how we get back to the tonic chord and our funky Gs :m chord becoming something else.