A tiny touch of soprano.
Inspired by much repeat listening of Light through the veins by John Hopkins. Not quite as subtle as his, might have to do some stretching + tuning of the sop samples to get something that fits the main chord progression. For now a little bit of whisper for tension and a little release to add to the ambient feel.